Relocate pkg/uuid to tempodb/backend
